requirements.txt
arabic-reshaper==2.1.0
asgiref==3.2.10
Babel==2.8.0
beautifulsoup4==4.9.1
braintree==4.3.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
console
cssselect==1.1.0
dj-database-url==0.5.0
Django==3.0.3
django-bootstrap-modal-forms==1.5.0
django-bootstrap3==12.1.0
django-bootstrap4==1.1.1
django-braces==1.14.0
django-db-file-storage==0.5.5
django-filter==2.0.0
django-formtools==2.2
django-heroku==0.3.1
django-payments==0.13.0
django-payments-razorpay==1.0
django-phone-field==1.8.1
django-sendsms==0.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
drfaddons==0.1.3
ezenv==0.91
Faker==4.0.3
future==0.18.2
gunicorn==19.9.0
html5lib==1.1
idna==2.9
Naked==0.1.31
parsel==1.6.0
paywix==1.3.0
phonenumbers==8.12.7
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2018.9
PyYAML==5.3.1
razorpay==1.2.0
reportlab==3.5.55
requests==2.23.0
shellescape==3.8.1
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
stripe==2.50.0
suds-jurko==0.6
text-unidecode==1.3
urllib3==1.25.9
w3lib==1.22.0
webencodings==0.5.1
whitenoise==4.1.2
wincertstore==0.2
xhtml2pdf==0.2.5
xmltodict==0.12.0